refactor(unity-ipc): split auth and method dispatch responsibilities
Why:
- Improve cohesion by separating request authorization concerns from IPC method-specific dispatching.

What Changed:
- Added IUnityIpcMethodDispatcher and UnityIpcMethodDispatcher for ping/execute/shutdown dispatch logic.
- Reduced UnityIpcRequestHandler responsibilities to session-token and protocol validation.
- Removed fallback dependency construction from UnityIpcServer so server focuses on lifecycle.
- Updated Unity daemon DI registration to resolve and inject method dispatcher.

Impact:
- Unity IPC server layer has clearer boundaries and smaller change surface per concern.
- Runtime behavior for authorized requests remains equivalent.
